Today I have done server side error handlin

+++ app.js
@@ -6,13 +6,27 @@ const path = require('path');
 const { url } = require('inspector');
 const methodOverRide = require('method-override');
 const ejsMate = require('ejs-mate');
+const wrapAsync = require('./utils/wrapAsync.js');
+const ExpressError = require('./utils/ExpressError.js');
+const listingValidationSchema = require('./schema.js');
 
 app.use(methodOverRide("_method"));
 app.set("view engine","ejs");
 app.set("views", path.join(__dirname,"views"));
 app.use(express.urlencoded({extended:true}));
+app.use(express.json());
 app.engine("ejs", ejsMate);
 app.use(express.static(path.join(__dirname,"/public")));
+
+const validateListing = (req,res,next)=>{
+  let {error} = listingValidationSchema.validate(req.body);
+  if(error){
+    let errMsg = error.details.map((el)=> el.message).join(",");
+    throw new ExpressError(400,errMsg);
+  }else{
+    next();
+  }
+}
 
 main().then(()=>{
     console.log("mongodb connected");
@@ -38,50 +52,56 @@ app.get("/listings/new",(req,res)=>{
 });
 
 //Route to add the new listing to database
-app.post("/listings",async (req,res)=>{
+app.post("/listings",validateListing, wrapAsync(async (req,res)=>{
     console.log("Request to post the data to DB is received");
+    
     let {title,description,price,url,location,country} = req.body;
-    await Listing.insertOne({
-        title:title,
-        description:description,
-        image:{
-            url:url
+     const listingData = {
+        title,
+        description,
+        image: {
+            url: url
         },
-        price:price,
-        location:location,
-        country:country
-    });
+        price,
+        location,
+        country
+    };
+
+    await Listing.create(listingData);
     
     res.redirect("/listings");
-});
+}));
 
 //Route to get all the listings
-app.get("/listings",async (req,res)=>{
+app.get("/listings",wrapAsync(async (req,res)=>{
     let listings =await Listing.find();
     console.log("request to all the listings is received");
     res.render("listings/index.ejs",{listings});
-});
+}));
 
 //Route to show the info of only one listng
-app.get("/listings/:id",async (req,res)=>{
+app.get("/listings/:id",wrapAsync(async (req,res)=>{
     let {id} = req.params;
     let details =await Listing.findById(id);
     console.log(`Request to show listing is received with id ${id}`);
     res.render("listings/show.ejs",{details});
-});
+}));
 
 //Route to edit the details
-app.get("/listings/:id/edit",async (req,res)=>{
+app.get("/listings/:id/edit",wrapAsync(async (req,res)=>{
     let {id} = req.params;
     let obj=await Listing.findById(id);
     console.log("Request to edit is received");
     res.render("listings/edit.ejs",{obj});
-});
+}));
 
 //Route to put the data edited data in DB
-app.put("/listings/:id",async (req,res)=>{
+app.put("/listings/:id",validateListing,wrapAsync(async (req,res)=>{
     let {id} = req.params;
     let {title,description,price,url,location,country} = req.body;
+     if(!title || !description || !price || !url || !location || !country){
+        throw new ExpressError(400,"All the fields are required!! Please send data");
+    }
     console.log("edits saved");
     await Listing.findByIdAndDelete(id);
     await Listing.insertOne({
@@ -95,14 +115,26 @@ app.put("/listings/:id",async (req,res)=>{
         country:country
     })
     res.redirect("/listings");
-});
+}));
 
 //Route to delete a listing
-app.delete("/listings/:id",async (req,res)=>{
+app.delete("/listings/:id",wrapAsync(async (req,res)=>{
     let {id} = req.params;
     await Listing.findByIdAndDelete(id);
     console.log("Request to delete a listing is received");
     res.redirect("/listings");
+}));
+
+//Error Handling route
+app.all(/.*/,(req,res,next)=>{   //Here I have used /.*/  ---It is a js regular expression and which is used for receiving requests from any route
+    next(new ExpressError(404,"page not found")); 
+});
+
+
+//Error Handling Middle-wares
+app.use((err,req,res,next)=>{
+    let {status=500 , message="Something went wrong"} = err;
+    res.status(status).render("error.ejs",{err});
 });
 
 app.listen(8080,()=>{
